requirements.txt
# Core video/audio processing
yt-dlp>=2024.12.0
gallery-dl>=1.27.0
ffmpeg-python>=0.2.0

# Audio analysis
librosa>=0.10.0
scipy>=1.11.0
numpy>=1.24.0
soundfile>=0.12.0

# CLI and UI
click>=8.1.0
rich>=13.0.0

# Utilities
requests>=2.31.0
validators>=0.22.0
python-dotenv>=1.0.0